refactor: change Ethereum naming
